Move Bluemix test to Travis CI cronjob
Previously, all events that were not pull requests triggered a Bluemix
deployment. This caused failures for all forks that did not have Travis
CI configured to deploy to Bluemix.

Now, we trigger our test that had been only for Pull Requests on all
events that are not cron events, and the Bluemix test to run only in
cronjobs.

**NOTE:** Work on getting other testing working is in
IBM/OpenWhisk-on-Kubernetes#21.

Related-Issue: IBM/OpenWhisk-on-Kubernetes#23
Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>